requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.5.0

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.1

# Image handling
Pillow==10.2.0

# Environment variables
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
django-extensions==3.2.3

# AI & Machine Learning
anthropic==0.75.0

# =============================================================================
# Production Dependencies
# =============================================================================

# Centralized Logging & Error Tracking
sentry-sdk==1.39.1

# Cloud Storage (AWS S3)
django-storages[boto3]==1.14.2
boto3==1.34.14

# Security (HMAC signature verification already in stdlib)

# WSGI Server for Production
gunicorn==21.2.0

# Static Files (WhiteNoise for serving static files)
whitenoise==6.6.0

# Password hashing (Argon2)
argon2-cffi==23.1.0
